Move logos out of ui-common
