--- ui-common/components/hemiLogo.tsx
@@ -1,46 +1,0 @@
-import { ComponentProps } from 'react'
-const HemiMainColor = '#FF5F00'
-
-export const HemiTokenWithBackground = (props: ComponentProps<'svg'>) => (
-  <svg
-    fill="none"
-    height={20}
-    viewBox="0 0 20 20"
-    width={20}
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}
-  >
-    <rect fill="#FFEBD4" height={20} rx={10} width={20} />
-    <path
-      d="M11.153 4.002a.112.112 0 0 0-.13.092l-.745 4.302h-.556l-.745-4.302a.112.112 0 0 0-.13-.092c-2.645.521-4.673 2.815-4.84 5.619 0 .004-.007.122-.007.181v.196c0 2.974 2.085 5.449 4.85 5.997.062.01.12-.03.13-.093l.746-4.302h.556l.74 4.306c.012.063.07.103.131.092 2.645-.525 4.669-2.819 4.84-5.622 0-.004.007-.123.007-.182v-.196c.004-2.974-2.082-5.449-4.847-5.996Z"
-      fill="#FF6C15"
-    />
-  </svg>
-)
-
-export const HemiLogoFull = () => (
-  <svg fill="none" height={24} width={70} xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M29.089 5.259v5.787l-.29-.097c.193-.839.587-1.452 1.18-1.839.594-.4 1.303-.6 2.13-.6 1.057 0 1.896.355 2.515 1.065.632.697.948 1.638.948 2.825V19h-2.051v-5.864c0-.942-.161-1.645-.484-2.11-.322-.464-.826-.696-1.51-.696-.748 0-1.341.245-1.78.735-.439.477-.658 1.168-.658 2.07V19h-2.051V5.259h2.05Zm8.131 8.612c0-1.07.2-2.013.6-2.825.413-.813.987-1.439 1.723-1.878.735-.438 1.593-.658 2.574-.658.838 0 1.6.194 2.284.581.696.387 1.251.98 1.664 1.78.413.787.639 1.762.677 2.923l.02.658h-7.374c.064.955.335 1.703.813 2.245.477.529 1.116.793 1.916.793.516 0 .98-.141 1.393-.425a2.55 2.55 0 0 0 .949-1.162l2.148.155a4.308 4.308 0 0 1-1.684 2.323c-.813.567-1.748.851-2.806.851-.98 0-1.839-.22-2.574-.658-.736-.439-1.31-1.064-1.723-1.877-.4-.813-.6-1.755-.6-2.826Zm7.374-.948c-.116-.916-.4-1.587-.852-2.013-.438-.438-.98-.658-1.625-.658-.749 0-1.355.239-1.82.716-.464.465-.754 1.116-.87 1.955h5.167Zm5.83-4.18.038 1.683c.246-.606.62-1.077 1.123-1.412.516-.336 1.116-.504 1.8-.504.735 0 1.361.187 1.877.562.516.374.884.896 1.103 1.567.232-.696.607-1.225 1.123-1.587.529-.361 1.18-.542 1.955-.542.683 0 1.27.155 1.76.465.504.297.884.735 1.142 1.316.272.58.407 1.284.407 2.11V19H60.7v-5.96c0-.904-.154-1.594-.464-2.072-.31-.49-.755-.735-1.335-.735-.478 0-.89.11-1.239.329-.335.22-.593.542-.774.968-.18.412-.271.916-.271 1.51V19h-1.935v-5.96c0-.904-.162-1.594-.484-2.072-.31-.49-.749-.735-1.316-.735-.478 0-.89.11-1.239.329-.336.22-.593.542-.774.968-.18.412-.271.916-.271 1.51V19h-2.052V8.743h1.878Zm16.91 0V19h-2.051V8.743h2.051Zm.058-3.465v1.994h-2.148V5.278h2.148Z"
-      fill="#000202"
-    />
-    <path
-      d="M12.09 2.019a.19.19 0 0 0-.222.155L10.61 9.402h-.94L8.412 2.174a.19.19 0 0 0-.221-.155C3.72 2.895.295 6.748.012 11.459c0 .006-.012.205-.012.304V12.093c0 4.996 3.524 9.154 8.197 10.073a.19.19 0 0 0 .22-.155l1.26-7.228h.939l1.252 7.234a.19.19 0 0 0 .221.156c4.47-.883 7.89-4.736 8.179-9.447 0-.006.012-.205.012-.304V12.092c.006-4.996-3.518-9.154-8.19-10.073Z"
-      fill="#FF6C15"
-    />
-  </svg>
-)
-
-export const HemiSymbol = () => (
-  <svg
-    enableBackground="new 0 0 1080 1080"
-    viewBox="0 0 1080 1080"
-    xmlSpace="preserve"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M436.2 1069.8c5.5 1.1 10.8-2.6 11.7-8.1L515 681.5h49.9l67 380.2c1 5.5 6.3 9.2 11.7 8.1 238-46.3 420.2-248.8 435.3-496.4v.1c0-.2.6-10.9.7-16.1 0-.6 0-1.2.1-1.7.1-2 .1-3.9.1-5.9v-2.2c0-2.5.1-5 .1-7.4v-.1c0-262.6-187.7-481.4-436.2-529.8-5.5-1.1-10.8 2.6-11.7 8.1l-67 380.1h-49.9l-67-380.2c-1-5.5-6.3-9.2-11.7-8.1C198.3 56.4 16.2 259 1 506.6v-.1c0 .2-.6 10.9-.7 16.1 0 .6 0 1.2-.1 1.7-.1 2-.1 3.9-.1 5.9v2.2c0 2.5-.1 5-.1 7.4v.1c0 262.7 187.7 481.5 436.2 529.9z"
-      fill={HemiMainColor}
-    />
-  </svg>
-)
